fix: tidy up contexts
