+++ src/react-menu3/contexts/EventHandlersContext.ts
@@ -1,0 +1,11 @@
+import { createContext } from "react";
+import { RMEvent } from "../utils";
+
+export interface EventHandlersContextType {
+  handleClose?: () => void;
+  handleClick: (event: RMEvent, checked: boolean) => void;
+}
+
+export const EventHandlersContext = createContext<EventHandlersContextType>({
+  handleClick: () => {},
+});
