+++ src/react-menu3/contexts/HoverItemContext.ts
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const HoverItemContext = createContext(undefined);
