+++ src/react-menu3/contexts/ItemSettingsContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+
+export const ItemSettingsContext = createContext<{ submenuCloseDelay: number; submenuOpenDelay: number }>({
+  submenuOpenDelay: 0,
+  submenuCloseDelay: 0,
+});
