+++ src/react-menu3/contexts/RadioGroupContext.ts
@@ -1,0 +1,9 @@
+import { createContext } from "react";
+import { EventHandler } from "../types";
+import { RadioChangeEvent } from "../components/MenuRadioGroup";
+
+export const RadioGroupContext = createContext<{
+  value?: any;
+  name?: string;
+  onRadioChange?: EventHandler<RadioChangeEvent>;
+}>({});
