+++ src/react-menu3/utils/constants.ts
@@ -1,17 +1,4 @@
-import { createContext, MutableRefObject, RefObject } from "react";
-import {
-  ControlledMenuProps,
-  EventHandler,
-  FocusPosition,
-  MenuDirection,
-  MenuOverflow,
-  MenuReposition,
-  MenuState,
-  MenuViewScroll,
-  RectElement,
-  TransitionFieldType,
-} from "../types";
-import { RadioChangeEvent } from "../components/MenuRadioGroup";
+import { MenuState } from "../types";
 
 export const menuContainerClass = "szh-menu-container";
 export const menuClass = "szh-menu";
@@ -24,30 +11,6 @@ export const menuGroupClass = "group";
 export const subMenuClass = "submenu";
 export const radioGroupClass = "radio-group";
 
-export const HoverItemContext = createContext(undefined);
-
-interface MenuListItemContextType {
-  isParentOpen?: boolean;
-  isSubmenuOpen?: boolean;
-  dispatch: (actionType: number, item: any, nextIndex: FocusPosition) => void;
-  updateItems: (item: any, isMounted?: boolean) => void;
-  setOpenSubmenuCount: (fn: (count: number) => number) => void;
-}
-
-export const MenuListItemContext = createContext<MenuListItemContextType>({
-  dispatch: () => {},
-  updateItems: () => {},
-  setOpenSubmenuCount: () => 0,
-});
-
-export const MenuListContext = createContext<{
-  overflow?: MenuOverflow;
-  overflowAmt?: number;
-  parentMenuRef?: MutableRefObject<any>;
-  parentDir?: MenuDirection;
-  reposSubmenu?: boolean;
-}>({});
-
 export interface RMEvent {
   value: any;
   syntheticEvent: any;
@@ -55,42 +18,6 @@ export interface RMEvent {
   name?: string;
   key?: string;
 }
-
-export const EventHandlersContext = createContext<{
-  handleClose?: () => void;
-  handleClick: (event: RMEvent, checked: boolean) => void;
-}>({
-  handleClick: () => {},
-});
-
-export const RadioGroupContext = createContext<{
-  value?: any;
-  name?: string;
-  onRadioChange?: EventHandler<RadioChangeEvent>;
-}>({});
-
-interface SettingsContextType extends ControlledMenuProps {
-  rootMenuRef?: MutableRefObject<any>;
-  rootAnchorRef?: MutableRefObject<any>;
-  scrollNodesRef: MutableRefObject<{ anchors?: Element[]; menu?: any }>;
-
-  initialMounted?: boolean;
-  unmountOnClose?: boolean;
-  transition?: TransitionFieldType;
-  transitionTimeout?: number;
-  boundingBoxRef?: RefObject<Element | RectElement>;
-  boundingBoxPadding?: string;
-  reposition?: MenuReposition;
-  viewScroll?: MenuViewScroll;
-}
-
-// FIXME hacking a default context in here is probably bad
-export const SettingsContext = createContext<SettingsContextType>({} as SettingsContextType);
-
-export const ItemSettingsContext = createContext<{ submenuCloseDelay: number; submenuOpenDelay: number }>({
-  submenuOpenDelay: 0,
-  submenuCloseDelay: 0,
-});
 
 export const Keys = Object.freeze({
   ENTER: "Enter",
